Loosen rubyzip dependency's version constraint
There is a [published security
vulnerability](https://github.com/rubyzip/rubyzip/issues/315) in rubyzip
that requires upgrading it, but the prior version constraint did not
allow updating minor versions.

This change raises the minimum version of rubyzip to include the
vulnerability patch and drops the patch version from the constraint to
allow minor version updates.